taxi_delivery/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
idna==2.7
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
